refactor: rename `ports` folder
