+++ src/main/java/io/pakland/mdas/githubstats/domain/repository/CommentRepository.java
@@ -1,4 +1,4 @@
-package io.pakland.mdas.githubstats.domain.ports;
+package io.pakland.mdas.githubstats.domain.repository;
 
 import io.pakland.mdas.githubstats.domain.model.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
