+++ src/main/java/io/pakland/mdas/githubstats/domain/repository/CommitRepository.java
@@ -1,4 +1,4 @@
-package io.pakland.mdas.githubstats.domain.ports;
+package io.pakland.mdas.githubstats.domain.repository;
 
 import io.pakland.mdas.githubstats.domain.model.Commit;
 import org.springframework.data.jpa.repository.JpaRepository;
