+++ src/main/java/io/pakland/mdas/githubstats/domain/repository/HistoricQueriesRepository.java
@@ -1,4 +1,4 @@
-package io.pakland.mdas.githubstats.domain.ports;
+package io.pakland.mdas.githubstats.domain.repository;
 
 import io.pakland.mdas.githubstats.domain.model.HistoricQueries;
 import org.springframework.data.jpa.repository.JpaRepository;
