+++ src/main/java/io/pakland/mdas/githubstats/domain/repository/PullRequestsRepository.java
@@ -1,4 +1,4 @@
-package io.pakland.mdas.githubstats.domain.ports;
+package io.pakland.mdas.githubstats.domain.repository;
 
 import io.pakland.mdas.githubstats.domain.model.PullRequest;
 import org.springframework.data.jpa.repository.JpaRepository;
