+++ src/main/java/io/pakland/mdas/githubstats/domain/repository/TeamRepository.java
@@ -1,4 +1,4 @@
-package io.pakland.mdas.githubstats.domain.ports;
+package io.pakland.mdas.githubstats.domain.repository;
 
 import io.pakland.mdas.githubstats.domain.model.Team;
 import org.springframework.data.jpa.repository.JpaRepository;
